feat: create smart-contracts for creating course
now there are 2 smart contracts in the ton folder: Factory and Seller.
Factory is responsible for creating new smart contracts (Seller, Customer).
Now when viewing your created courses
there is a button to activate course and when you click on it,
ton wallet is opened and Seller smart contract is deployed in TON blockchain