fix: fix parentId problem
a request with an empty parentId was sent to the backend
and because of this an error was triggered on receiving all modules for 1 course.
Also removed React.StrictMode because it was sending 2x more requests to the backend.
Removed the updatePageData function.
For Select and Textarea, the border is now highlighted with the Telegram color instead of blue