feat: make the MarketplaceFee smart-contract work
now from each course activation on the blockchain (when created)
and from each purchase on the MarketplaceFee contract comes 0.6 TON.
Further 0.25 ton are sent to developers.

Add encodeTransferToMarketplace function.
It takes data from build, encodes it and then turns it into a cell.
his is needed to send payload message in transaction for MarketplaceFee contract.

If you have to change the MarketplaceFee smart contract in the future,
change the REACT_APP_MARKETPLACE_FEE_CONTRACT_OPCODE field in the .env file