feat: change contracts
Have a problem with the fact that at the moment
we can't get Purchase smart contract data via Course smart contract.
Since the Purchase contract is deployed via Course,
just fromInit (Purchase data) returns an incorrect contract address.
So it was necessary to get the data from the get function,
but these attempts were unsuccessful.
So it was decided to change the smart contract data
so that we could freely get the addresses of Course and Purchase contracts.