feat: add prodWallet1 and prodWallet2 to course smart-contract for marketplaceFee
Now it is not necessary to send a "Sale" message to the Course contract
but a message with the isProduction true or false field.
The testnet or mainnet addresses for marketplaceFee depend on it.
Also added the ability to withdraw funds from smart contracts for course and purchase.
Fixed tests for smart-contracts