feat: remove smart contract balance check
now as soon as a course is activated on the blockchain (Creating a Course smart contract)
or purchasing a course (Creating a smart contract Purchase)
a request is made to the backend
and checks the contract balance every 20 seconds for 5 min.
If the balance of the contract is increased,
then points are credited to the user and a notification is sent to the user.
If not increased, an error notification is sent to the user and no points are awarded.
This is done so that the user does not have to wait with a locked screen for the smart contract balance to be updated.
Since if the user would close the app, the points would not be credited.
The backend now performs this procedure in the background, unlike the frontend.

Also, there are now no requests (endpoints) to the backend to add points, this is done in the background