feat: self-appearing and self-correcting bug
The Purchase contract was shown with a balance of 0,
sent to the backend and since the balance of this contract did not change for some reason,
the points were not added and there was no notification of a successful purchase.
In transactions of the wallet contract another Purchase contract with courseId and other data with balance was used.
For some reason the contract that was shown in the interface and sent to the backend had 0 balance.
And the one in transactions - with a real positive balance.
After a lot of trying to solve it, it seems to have solved itself, not sure if my code had any effect on the fix.
And the most surprising thing is that the balances the user has now seem to have increased
since the other addresses were shown.
What that situation was, I haven't figured out yet