fix: fix webpack issue
If we load components using the lazy library (from react),
we need a plugin to filter errors related to the order in which css modules are loaded.
Adding this plugin causes errors with Buffer webpack.
After several hours of fixing this error, I had to remove the plugin for filtering errors,
make the command "npm update" and the error was eliminated