feat: add disabling user inputs after successful sending of data to the backend
now after clicking on the "Send confirmation code to email" button
and successfully saving the data in the database
there is no need to allow the user to change the data,
as new data after clicking on "Send" or "Send code again" buttons will not be sent
and it will confuse the user.
That is why the inputs are disabled.
Also now if a user does not have a saved phone number in the database,
an empty string is shown instead of null